Enhancement: move components and variables to directory outside of app/
