+++ src/variables/people.ts
