+++ src/variables/teams.ts
